chore(react): move chat components into subdir
