+++ packages/react/src/chat/ChatModal/ChatModal.test.tsx
@@ -1,7 +1,7 @@
 import { DrAliceSmith, MockClient } from '@medplum/mock';
 import { MedplumProvider } from '@medplum/react-hooks';
 import { MemoryRouter } from 'react-router-dom';
-import { act, fireEvent, render, screen } from '../test-utils/render';
+import { act, fireEvent, render, screen } from '../../test-utils/render';
 import { ChatModal, ChatModalProps } from './ChatModal';
 
 describe('ChatModal', () => {
