+++ packages/react/src/chat/ThreadChat/ThreadChat.test.tsx
@@ -5,7 +5,7 @@ import { BartSimpson, DrAliceSmith, HomerSimpson, MockClient } from '@medplum/mo
 import { MedplumProvider, _subscriptionController } from '@medplum/react-hooks';
 import crypto from 'node:crypto';
 import { MemoryRouter } from 'react-router-dom';
-import { act, fireEvent, render, screen } from '../test-utils/render';
+import { act, fireEvent, render, screen } from '../../test-utils/render';
 import { ThreadChat, ThreadChatProps } from './ThreadChat';
 
 type SubscriptionControllerEvents = {
