+++ packages/react/src/index.ts
@@ -16,9 +16,7 @@ export * from './AttachmentDisplay/AttachmentDisplay';
 export * from './AttachmentInput/AttachmentInput';
 export * from './BackboneElementDisplay/BackboneElementDisplay';
 export * from './BackboneElementInput/BackboneElementInput';
-export * from './BaseChat/BaseChat';
 export * from './CalendarInput/CalendarInput';
-export * from './ChatModal/ChatModal';
 export * from './CheckboxFormSection/CheckboxFormSection';
 export * from './CodeInput/CodeInput';
 export * from './CodeableConceptDisplay/CodeableConceptDisplay';
@@ -94,12 +92,14 @@ export * from './SearchFilterEditor/SearchFilterEditor';
 export * from './ServiceRequestTimeline/ServiceRequestTimeline';
 export * from './SmartAppLaunchLink/SmartAppLaunchLink';
 export * from './StatusBadge/StatusBadge';
-export * from './ThreadChat/ThreadChat';
 export * from './Timeline/Timeline';
 export * from './TimingInput/TimingInput';
 export * from './ValueSetAutocomplete/ValueSetAutocomplete';
 export * from './auth/RegisterForm';
 export * from './auth/SignInForm';
+export * from './chat/BaseChat/BaseChat';
+export * from './chat/ChatModal/ChatModal';
+export * from './chat/ThreadChat/ThreadChat';
 export * from './utils/date';
 export * from './utils/outcomes';
 export * from './utils/questionnaire';
